requirements.txt
discord.py
discord.py[voice]
apscheduler
aiofiles
asyncio
loguru
psycopg2
psycopg2-binary
jishaku
buttons
pillow
coloredlogs
requests
wavelink
better-profanity==0.6.1
psutil
transliterate
python-dotenv
git+https://github.com/Rapptz/discord-ext-menus
